requirements.txt
DateTime==4.4
feedparser==6.0.8
python-dateutil==2.8.2
pytz==2022.1
sgmllib3k==1.0.0
six==1.16.0
zope.interface==5.4.0
